requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-WTF==1.1.1

# Scientific Computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
pandas>=2.0.0

# Computer Vision
opencv-python-headless>=4.7.0
Pillow>=9.5.0
scikit-image>=0.20.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.2.0

# Utilities
requests>=2.28.0
jinja2>=3.1.0
markupsafe>=2.0.0
tqdm>=4.65.0

# Transformers and ML (install from PyPI)
transformers>=4.35.0
timm>=0.9.0
accelerate>=0.24.0
tokenizers>=0.14.0
datasets>=2.5.0

# PyTorch (CPU-only) — from PyTorch's custom index
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
torchaudio==2.1.0+cpu
